+++ src/webserver/webserver.c
@@ -744,10 +744,7 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
                 if (mg_user_data->mympd_api_started == false) {
                     //mympd_api thread not yet ready
                     MYMPD_LOG_WARN(frontend_nc_data->partition, "mympd_api thread not yet ready");
-                    sds response = jsonrpc_respond_message(sdsempty(), GENERAL_API_NOT_READY, 0,
-                        JSONRPC_FACILITY_GENERAL, JSONRPC_SEVERITY_ERROR, "myMPD not yet ready");
-                    webserver_send_data(nc, response, sdslen(response), EXTRA_HEADERS_JSON_CONTENT);
-                    FREE_SDS(response);
+                    webserver_send_jsonrpc_response(nc, GENERAL_API_NOT_READY, 0, JSONRPC_FACILITY_GENERAL, JSONRPC_SEVERITY_ERROR, "myMPD not yet ready");
                 }
                 //check partition
                 if (get_partition_from_uri(nc, hm, frontend_nc_data) == false) {
@@ -764,10 +761,7 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
                 FREE_SDS(body);
                 if (rc == false) {
                     MYMPD_LOG_ERROR(frontend_nc_data->partition, "Invalid API request");
-                    sds response = jsonrpc_respond_message(sdsempty(), GENERAL_API_UNKNOWN, 0,
-                        JSONRPC_FACILITY_GENERAL, JSONRPC_SEVERITY_ERROR, "Invalid API request");
-                    webserver_send_data(nc, response, sdslen(response), EXTRA_HEADERS_JSON_CONTENT);
-                    FREE_SDS(response);
+                    webserver_send_jsonrpc_response(nc, GENERAL_API_UNKNOWN, 0, JSONRPC_FACILITY_GENERAL, JSONRPC_SEVERITY_ERROR, "Invalid API request");
                 }
             }
             else if (mg_match(hm->uri, mg_str("/albumart-thumb/*"), NULL)) {
@@ -834,6 +828,10 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
             }
         #ifdef MYMPD_ENABLE_LUA
             else if (mg_match(hm->uri, mg_str("/script-api/*"), NULL)) {
+                if (config->scripts_external == false) {
+                    MYMPD_LOG_ERROR(frontend_nc_data->partition, "External scripts are disabled");
+                    webserver_send_jsonrpc_response(nc, GENERAL_API_UNKNOWN, 0, JSONRPC_FACILITY_SCRIPT, JSONRPC_SEVERITY_ERROR, "External scripts are disabled");
+                }
                 //enforce script acl
                 if (enforce_acl(nc, config->scriptacl) == false) {
                     break;
@@ -847,10 +845,7 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
                 FREE_SDS(body);
                 if (rc == false) {
                     MYMPD_LOG_ERROR(frontend_nc_data->partition, "Invalid script API request");
-                    sds response = jsonrpc_respond_message(sdsempty(), GENERAL_API_UNKNOWN, 0,
-                        JSONRPC_FACILITY_SCRIPT, JSONRPC_SEVERITY_ERROR, "Invalid script API request");
-                    webserver_send_data(nc, response, sdslen(response), EXTRA_HEADERS_JSON_CONTENT);
-                    FREE_SDS(response);
+                    webserver_send_jsonrpc_response(nc, GENERAL_API_UNKNOWN, 0, JSONRPC_FACILITY_SCRIPT, JSONRPC_SEVERITY_ERROR, "Invalid script API request");
                 }
             }
             else if (mg_match(hm->uri, mg_str("/script/*/*"), NULL)) {